+++ c++/utils/hash_utils.cpp
@@ -6,7 +6,7 @@
 
 
 //SHA-256
-void sha256_string(const unsigned char *message, unsigned char *output_buffer, unsigned long long mlen) {
+void sha256_string(const unsigned char *message, unsigned char *output_buffer, size_t mlen) {
     SHA256_CTX sha256;
     SHA256_Init(&sha256);
     SHA256_Update(&sha256, message, mlen);
@@ -32,7 +32,7 @@ void sha256_file(const char *file_path, unsigned char *output_buffer) {
 }
 
 //SHA-384
-void sha384_string(const unsigned char *message, unsigned char *output_buffer, unsigned long long mlen) {
+void sha384_string(const unsigned char *message, unsigned char *output_buffer, size_t mlen) {
     SHA512_CTX sha384;
     SHA384_Init(&sha384);
     SHA384_Update(&sha384, message, mlen);
@@ -58,7 +58,7 @@ void sha384_file(const char *file_path, unsigned char *output_buffer) {
 }
 
 //SHA-512
-void sha512_string(const unsigned char *message, unsigned char *output_buffer, unsigned long long mlen) {
+void sha512_string(const unsigned char *message, unsigned char *output_buffer, size_t mlen) {
     SHA512_CTX sha512;
     SHA512_Init(&sha512);
     SHA512_Update(&sha512, message, mlen);
